[BE] [feat] #40 : 펀딩 필드 변경 및 삭제 기능 구현

See merge request s11-fintech-finance-sub1/S11P21E202!52

+++ backend/givutake/src/main/java/com/accepted/givutake/funding/entity/FundingParticipants.java
@@ -19,7 +19,7 @@ public class FundingParticipants extends BaseTimeEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "participant_idx", nullable = false)
-    private int participantIdx;
+    private long participantIdx;
 
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "funding_idx", nullable = false)
